Feature: Add support for `.xctestproducts` and Test Plan (#606)
<!-- Please provide brief information about the PR, what it contains &
its purpose, new behaviors after the change. And let us know here if you
need any help: https://github.com/microsoft/HydraLab/issues/new -->

## Description
In this PR, I suggest adding support for the `.xctestproducts` package
in XCTest. As I described in issue #583, using `xctestproducts` is more
suitable for CI and also supports the Test Plan feature.

Currently, in order to export the test products, one needs to package
the Derived Data and the `.xctestrun` file. After this change, the user
can choose to upload the `xctestproducts` package, and we will detect it
internally.

One additional change on the UI side is regarding the Test Plan name. In
case the test type is XCTest, the user can enter the test plan name,
which we will use if the test package is of type `xctestproducts`.

If we decide to proceed with this change, the [Wiki
page](https://github.com/microsoft/HydraLab/wiki/Create-test-build-and-run-XCTest)
will need to be updated.

### Linked GitHub issue ID: #583 

## Pull Request Checklist
<!-- Put an x in the boxes that apply. This is simply a reminder of what
we are going to look for before merging your code. -->

- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Code compiles correctly with all tests are passed.
- [x] I've read the [contributing
guide](https://github.com/microsoft/HydraLab/blob/main/CONTRIBUTING.md#making-changes-to-the-code)
and followed the recommended practices.
- [ ] [Wikis](https://github.com/microsoft/HydraLab/wiki) or
[README](https://github.com/microsoft/HydraLab/blob/main/README.md) have
been reviewed and added / updated if needed (for bug fixes / features)

### Does this introduce a breaking change?
*If this introduces a breaking change for Hydra Lab users, please
describe the impact and migration path.*

- [ ] Yes
- [x] No

## How you tested it
I test the change locally by uploading `xctestproducts` package and by
choosing different test plans to execute


Please check the type of change your PR introduces:
- [ ] Bugfix
- [x] Feature
- [ ] Technical design
- [ ] Build related changes
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Code style update (formatting, renaming) or Documentation content
changes
- [ ] Other (please describe):